% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE.R
\name{ATE_est}
\alias{ATE_est}
\title{causal dose-response curve}
\usage{
ATE_est(fY, fw, fA)
}
\arguments{
\item{fY}{outcome}

\item{fw}{balance weights}

\item{fA}{a continuous treatment}
}
\value{
tt_ATE1 is a linear dose-response function estimated using IPTW method.
}
\description{
causal dose-response curve
}
