% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goaldata.data2}
\alias{goaldata.data2}
\title{goaldata.data2}
\format{
A \code{\link[=data.frame]{data.frame()}} with two columns (x,y) representing some points in Greece.
}
\source{
Dimitris Kavroudakis
}
\description{
A  data2:
\itemize{
\item \strong{x}: The x coordinates of the points
\item \strong{y}: The y coordinates of the points
}
}
\examples{
library(goal)
data(goaldata.data2)
}
