% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goaldata.GRregions}
\alias{goaldata.GRregions}
\title{The 13 regions of Greece}
\format{
A \code{\link[=SpatialPolygonsDataFrame]{SpatialPolygonsDataFrame()}}  of the 13 regions of Greece
}
\source{
The shapefile of the corresponding polygons is available from
the \href{http://geodata.gov.gr/geodata}{Open-Data portal} of
the Greek Government.
Population and GDP values are available from
the
\href{http://epp.eurostat.ec.europa.eu/portal/page/portal/eurostat/home/}{Eurostat}
}
\description{
A \link{SpatialPolygonsDataFrame}  of the 13 regions of Greece.
The variables in the dataset are as follows:
\itemize{
\item \strong{code}: The ID of the region
\item \strong{Name_en}: Name of the region
\item \strong{gdp_capita}: The GDP per capita of the region
\item \strong{GDP_mil}: The GDP of the region in milion euros
\item \strong{pop_dens}: Population density of the region
\item \strong{pop}: Population of the region
\item \strong{total_area}: Total area of the region in \eqn{km^2}
}
}
\examples{
library(goal)
data(goaldata.GRregions)
}
