% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goaldata.roads}
\alias{goaldata.roads}
\title{A dataset of 40 roads}
\format{
A \code{\link[=data.frame]{data.frame()}} with 40 rows and 6 variables
}
\source{
\href{http://www.dimitrisk.gr}{Dimitris Kavroudakis}
}
\description{
The variables in the dataset are as follows:
\itemize{
\item \strong{from}: The id of the origin
\item \strong{to}: The id of the destination
\item \strong{id}: The id of this road
\item \strong{name}: The name of this road
\item \strong{cost}: The cost units of using this road
\item \strong{category}: The category of this road
}
}
\examples{
library(goal)
data(goaldata.roads)
}
