% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goaldata.towns}
\alias{goaldata.towns}
\title{A dataset of 15 towns}
\format{
A \code{\link[=data.frame]{data.frame()}} with 15 rows and 4 variables
}
\source{
\href{http://www.dimitrisk.gr}{Dimitris Kavroudakis}
}
\description{
The variables in the dataset are as follows:
\itemize{
\item \strong{id}: The id of the town
\item \strong{name}: The name of the town
\item \strong{pop}: The population of the town
\item \strong{meanincome}: The meanincome of the town
}
}
\examples{
library(goal)
data(goaldata.towns)
}
