% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.bb_2_pol}
\alias{osm.bb_2_pol}
\title{osm.bb_2_pol}
\usage{
osm.bb_2_pol(
  inVec = c(26.545029, 39.088569, 26.570177, 39.11681),
  outcrs = 4326
)
}
\arguments{
\item{inVec}{Four}

\item{outcrs}{CRS of the the output}
}
\value{
An sf polygon
}
\description{
Bounding Box to Polygon
Either by name of area (inPerioxi) or by bounding box (incoordinates).
}
\examples{
library(goal)
library(sf)

q=c(26.545029,39.088569,26.570177,39.116810)
poly = goal::osm.bb_2_pol(q, outcrs=4326)

}
\seealso{
Other osm: 
\code{\link{osm.ClipSFnetwork_with_poly}()},
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getLength_footway}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.myBB_2_pol}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{bbox,}
\keyword{network}
\keyword{openstreetmap,}
