% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.ClipSFnetwork_with_poly}
\alias{osm.ClipSFnetwork_with_poly}
\title{osm.ClipSFnetwork_with_poly}
\usage{
osm.ClipSFnetwork_with_poly(innet, inpol)
}
\arguments{
\item{innet}{sfnetwork to clip}

\item{inpol}{sf polygon to use for clipping}
}
\value{
An sfnetwork
}
\description{
Clip sfnetwork by Polygon
}
\examples{
library(goal)
library(sf)
library(sfnetworks)
library(tidygraph)

q=c(26.545029,39.088569,26.570177,39.116810)
net2 = osm.getRoads(q, withBB=TRUE, outcrs=4326)
poly = osm.bb_2_pol(q, outcrs =  4326)

net3 = osm.ClipSFnetwork_with_poly(net2, poly)

plot(net3, col="grey")
plot(poly, add = TRUE)
}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.bb_2_pol}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getLength_footway}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{bbox,}
\keyword{network,}
\keyword{openstreetmap,}
\keyword{polygon}
