% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.combineShops}
\alias{osm.combineShops}
\title{osm.combineShops}
\usage{
osm.combineShops(inam)
}
\arguments{
\item{inam}{a}
}
\value{
A nice
}
\description{
Combine all Shops results into single dataset. It merges \code{Points}, \code{Polygons} and \code{Myltipolygons} into a
single dataset of Points using centroids of polygons when necessary.
}
\examples{
library(goal)
sh = osm.getPOI(inPerioxi = "Mytilene Municipal Unit", inkey = "shop")
sh
shops = osm.combineShops(sh)
shops
}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.getClipedRoads}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{POIs}
\keyword{frequency,}
\keyword{openstreetmap,}
