% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getClipedRoads}
\alias{osm.getClipedRoads}
\title{osm.getClipedRoads}
\usage{
osm.getClipedRoads(
  incoordinates = c(26.545029, 39.088569, 26.570177, 39.11681),
  outcrs = 2100
)
}
\arguments{
\item{incoordinates}{An}

\item{outcrs}{An}
}
\value{
An sfnetwork
}
\description{
Get the roads of an area by bounding box.
}
\examples{
library(goal)
library(sf)
library(sfnetworks)
library(osmdata)
mynetwork = osm.getClipedRoads(incoordinates= c(26.545029,39.088569,26.570177,39.116810), 
  outcrs=2100
)
mynetwork 
 
}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{bbox,}
\keyword{network}
\keyword{openstreetmap,}
