% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getLength_footway}
\alias{osm.getLength_footway}
\title{osm.getLength_footway}
\usage{
osm.getLength_footway(place = "Mytilene Municipal Unit")
}
\arguments{
\item{place}{The name of an area.}
}
\value{
The total length of  footpaths in this area. This may be considered as a proxy for walcability score of this area.
}
\description{
Get length of footway for an area. This is a proxy of walkability score.
we get back a number.
}
\examples{
library(goal)
mylength = osm.getLength_footway( place="Mytilene Municipal Unit" )
mylength
}
\seealso{
Other osm: 
\code{\link{osm.ClipSFnetwork_with_poly}()},
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.bb_2_pol}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.myBB_2_pol}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{POIs}
\keyword{footway,}
\keyword{openstreetmap,}
\keyword{road,}
\keyword{walkability,}
