% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getPOI}
\alias{osm.getPOI}
\title{osm.getPOI}
\usage{
osm.getPOI(inPerioxi = "Mytilene Municipal Unit", inkey = "amenity")
}
\arguments{
\item{inPerioxi}{a perioxi}

\item{inkey}{name of the key}
}
\value{
A nice
}
\description{
Get all POIs of a city. Providing just the key (e.x. amenities, shop, ...) we get back an osmdata object
of Spatial Feature ("sf") type.
}
\examples{
library(goal)

}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getRoads}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{POIs}
\keyword{frequency,}
\keyword{openstreetmap,}
