% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getPOI_usingbb}
\alias{osm.getPOI_usingbb}
\title{osm.getPOI_usingbb}
\usage{
osm.getPOI_usingbb(
  inbb = c(26.547303, 39.101658, 26.564641, 39.113247),
  inkey = "amenity"
)
}
\arguments{
\item{inbb}{A vector of coordinates, representing the bounding box of the area.}

\item{inkey}{The name of a key such as: \code{amenities}, \code{shops}, ...}
}
\value{
An \code{osmdata} object of Spatial Feature ("sf") type.
}
\description{
Get all POIs of a city. Providing the bounding box
we get back an osmdata object of Spatial Feature ("sf") type.
}
\examples{
library(goal)
myt = osm.getPOI_usingbb( c(26.547303,39.101658,26.564641,39.113247) )
myt
}
\seealso{
Other osm: 
\code{\link{osm.ClipSFnetwork_with_poly}()},
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.bb_2_pol}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getLength_footway}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{POIs}
\keyword{frequency,}
\keyword{openstreetmap,}
