% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getRoads}
\alias{osm.getRoads}
\title{osm.getRoads}
\usage{
osm.getRoads(inPerioxi = "Mytilene Municipal Unit")
}
\arguments{
\item{inPerioxi}{ssd}
}
\value{
A nice
}
\description{
Get roads of a city
of Spatial Feature ("sf") type.
}
\examples{
library(goal)

}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{network}
\keyword{openstreetmap,}
