% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.myBB_2_pol}
\alias{osm.myBB_2_pol}
\title{osm.myBB_2_pol}
\usage{
osm.myBB_2_pol(n, s, w, e, bbox = NA, proj4string = sp::CRS("+init=epsg:4326"))
}
\arguments{
\item{n}{the top north latitude}

\item{s}{the bottom south latitude}

\item{w}{the most western longitude}

\item{e}{the most eastern longitude}

\item{bbox}{either a bounding box}

\item{proj4string}{either a projection}
}
\value{
An polygon
}
\description{
Create a polygon from a BB object
}
\examples{
library(devtools)
#install_github("dimitrisk/goal")
library(raster)
library(goal)
bb = c(38.1 , 26.5 , 39.05 , 27.7)
result = goal::osm.myBB_2_pol(n =bb[3], s =bb[1], w =bb[4], e =bb[2])
result
}
\seealso{
Other osm: 
\code{\link{osm.ClipSFnetwork_with_poly}()},
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.bb_2_pol}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getLength_footway}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getRoads}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{openstreetmap}
