% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.osmdata_result_2_bbox_pol}
\alias{osm.osmdata_result_2_bbox_pol}
\title{osm.osmdata_result_2_bbox_pol}
\usage{
osm.osmdata_result_2_bbox_pol(osmdata_result)
}
\arguments{
\item{osmdata_result}{An \code{osmdata} result.}
}
\value{
An SF polygon which is the bbox of the \code{osmdata} result.
}
\description{
Get the bbox of a result, as an SF polygon.
}
\examples{
library(devtools)
#install_github("dimitrisk/goal")
library(goal)
library(sf)
library(stringr)
am = goal::osm.getPOI_usingbb( c(26.547303,39.101658,26.564641,39.113247), inkey ="amenity" )
amenities = goal::osm.combineAmenities(am)
thepol = goal::osm.osmdata_result_2_bbox_pol(am)

}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{bbox,}
\keyword{openstreetmap,}
\keyword{polygon}
