% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.r
\name{spa.MeanCenter}
\alias{spa.MeanCenter}
\title{spa.MeanCenter}
\usage{
spa.MeanCenter(indf)
}
\arguments{
\item{indf}{A n}
}
\value{
A
}
\description{
Generate points along lines. Following a geometrical line, it produces points based on a factor.
}
\details{
Generates a number of points along lines.
}
\examples{
library(goal)
co = data.frame(x = c(0,3.9,4.1,8), y = c(1,4.1,3.9,7))
names(co) = c("x","y")
(co)

goal::spa.MeanCenter(indf=co)


}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{spa}
\keyword{point_pattern_analysis}
\keyword{spatial_analysis}
