% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uav.dimkaAlgorithm.r
\name{uav.DoReduction}
\alias{uav.DoReduction}
\title{uav.DoReduction}
\usage{
uav.DoReduction(intargets, insize = 15)
}
\arguments{
\item{intargets}{SpatialPointsDataFrame with the location of targets to reduce}

\item{insize}{not used so far.}
}
\value{
A list with polygons, points and sminos with the reduced target locations.
}
\description{
Do the reduction of targets.
}
\examples{
library(goal)
data(goaldata.data2)
one = goal::uav.GenerateTargets(goaldata.data2)
result = uav.DoReduction(one, insize=20)
print(result)
}
