% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uav.dimkaAlgorithm.r
\name{uav.GenerateTargets}
\alias{uav.GenerateTargets}
\title{uav.GenerateTargets}
\usage{
uav.GenerateTargets(
  data2,
  SourceCRS = "+init=epsg:4326",
  TargetCRS = "+init=epsg:2100",
  ploting = 0,
  export = 0,
  dosimplify = 0,
  output_folder = "Results"
)
}
\arguments{
\item{data2}{points of the polygon study area.}

\item{SourceCRS}{Source CRS of the points}

\item{TargetCRS}{Target CRS of the points}

\item{ploting}{Boolean whether to plot the results.}

\item{export}{Boolean}

\item{dosimplify}{Boolean whether to simplify the geometry of Polygon of the study area.}

\item{output_folder}{Path of the export folder}
}
\value{
A list of three objects:
\itemize{
\item \strong{Polygon}: A \code{\link[=SpatialPolygonsDataFrame]{SpatialPolygonsDataFrame()}} of the ConvexHull polygon of the area
\item \strong{simia}: A \code{\link[=SpatialPointsDataFrame]{SpatialPointsDataFrame()}} of the targets.
\item \strong{sminos}: A \code{\link[=SpatialPointsDataFrame]{SpatialPointsDataFrame()}} of the initial data points
\item \strong{lines}: A \code{\link[=SpatialLinesDataFrame]{SpatialLinesDataFrame()}} of the lines used for the calculation
}
}
\description{
Generate locations for targets (points) inside the study area (polygon) keeping low density of points,
covering all parts of the area and following the shape of the study area.
}
\examples{
library(goal)
data(goaldata.data2, package ="goal")
library(sp)
one = goal::uav.GenerateTargets(goaldata.data2)
plot(one$Polygon)
plot(one$sminos, col="red", add=TRUE)
plot(one$simia, col="blue", add=TRUE)
}
