% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uav.dimkaAlgorithm.r
\name{uav.removeOnePointByDM}
\alias{uav.removeOnePointByDM}
\title{uav.removeOnePointByDM}
\usage{
uav.removeOnePointByDM(insimia)
}
\arguments{
\item{insimia}{something}
}
\value{
a SpatialPointsDataFrame with 1 reduced point
}
\description{
Eliminate one point by Distance Matrix of points
}
\keyword{internal}
