% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uav.dimkaAlgorithm.r
\name{uav.showFocal}
\alias{uav.showFocal}
\title{uav.showFocal}
\usage{
uav.showFocal(insimia, inpol, size = 20, toplot = TRUE)
}
\arguments{
\item{insimia}{something}

\item{inpol}{a simia SpatialPOintsDataFrame}

\item{size}{the cell size (rows-columns)}

\item{toplot}{Whether to plot a diagram of the focal density of points in the study area}
}
\value{
a frequency table of number of values per cell of the focal operation
}
\description{
Show the Focal map of the study area.
}
\examples{
library(goal)
data(goaldata.data2, package ="goal")
one = goal::uav.GenerateTargets(goaldata.data2)
result = goal::uav.showFocal(one$simia, one$Polygon, size=20, toplot=TRUE)
print(result)
}
