% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis.BarPlot}
\alias{vis.BarPlot}
\title{vis.BarPlot}
\usage{
vis.BarPlot(
  inTable,
  percentage = F,
  subtitle = TRUE,
  main = "main",
  ylab = "",
  decimals = 1
)
}
\arguments{
\item{inTable}{The table to plot}

\item{percentage}{Whether to use percentages of absolute numbers}

\item{subtitle}{Whether to add an informative subtitle in the barplot}

\item{main}{The main title of the barplot}

\item{ylab}{The title in the \code{Y} axes}

\item{decimals}{Number of decimals to use}
}
\value{
A nice barplot of the \code{inTable}
}
\description{
A nice barplot of a \link{table} with bar labels and absolute/relative frequencies.
}
\examples{
library(goal)
library(plotrix)
cars = table( c(1,1,1,1,3,2,4,3,5,4,6,5,1,5,5,2,2,2,3,3,3,1,1,2) )
vis.BarPlot(inTable = cars, percentage = TRUE,  subtitle = TRUE,
main = "Relative frequency of cars", ylab = "Cars", decimals=1)

}
\seealso{
Other vis: 
\code{\link{vis.Hist}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{vis}
\keyword{visualization}
