% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis.Hist}
\alias{vis.Hist}
\title{vis.Hist}
\usage{
vis.Hist(inVector, ecdf = T, ...)
}
\arguments{
\item{inVector}{The vector of numbers to plot}

\item{ecdf}{Boolean for the addition of the red line of the Cumulative Density at the right of the histogram plot}

\item{...}{Additional arguments to be passed to the 'hist' command of the base histogram.}
}
\value{
A nice histogram of a single variable with an additional Cumulative Density curve.
}
\description{
A nice histogram of a single variable with an additional Cumulative Density curve.
}
\examples{
library(goal)

set.seed(1821)
myNumbers = rnorm(500, 130, 10)
vis.Hist(myNumbers, freq=FALSE, main="This is the main title", xlab="X axis title")

}
\seealso{
Other vis: 
\code{\link{vis.BarPlot}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{vis}
\keyword{visualization}
