% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWC.R
\name{wAMD_function}
\alias{wAMD_function}
\title{dual-weigt correlation (DWC) for selecting optimal lambda}
\usage{
wAMD_function(DataM, varlist, trt.var, wgt, beta)
}
\arguments{
\item{DataM}{Data that include outcome, treatment and all of covariates}

\item{varlist}{names of potential confounders including known and unknown}

\item{trt.var}{names of treatment}

\item{wgt}{balancing weights}

\item{beta}{the coefficient of covariates in the unpenalized "full" outcome model}
}
\value{
wAMD is a vector of dual-weight correlation.
}
\description{
dual-weigt correlation (DWC) for selecting optimal lambda
}
