% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOAL.R
\name{GOAL}
\alias{GOAL}
\title{Genelized outcome-adaptive LASSO estimation}
\usage{
GOAL(
  data,
  var.list,
  covar = NULL,
  Trt = "Trt",
  out = "Y",
  lambda_vec = c(-10, -5, -2, -1, -0.75, -0.5, -0.25, 0.25, 0.49),
  gamma_convergence = 2
)
}
\arguments{
\item{data}{Data that include outcome, treatment and all of covariates}

\item{var.list}{names of potential unknown confounders}

\item{covar}{names of known confounders}

\item{Trt}{names of treatment}

\item{out}{names of outcome}

\item{lambda_vec}{a vector of possible lambda values and defualt is c( -10, -5, -2, -1, -0.75, -0.5, -0.25, 0.25, 0.49)}

\item{gamma_convergence}{the value of gamma converagence and defualt is 2}
}
\value{
ATE is the linear dose-response function estimated using IPTW method.    selectedVar is covariates that GOAL selected. Ideally, it include confounders and prognostic covariates.    lambda is the optimal lambda corresponding to minimized DWC.
}
\description{
Genelized outcome-adaptive LASSO estimation
}
