% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.combineAmenities}
\alias{osm.combineAmenities}
\title{osm.combineAmenities}
\usage{
osm.combineAmenities(inam)
}
\arguments{
\item{inam}{Amenities}
}
\value{
A nice
}
\description{
Combine all AMENITIES results into single dataset (Points, Polygons, Myltipolygons) (centroids).
}
\examples{
library(goal)

}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{POIs}
\keyword{frequency,}
\keyword{openstreetmap,}
