% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getFrequency}
\alias{osm.getFrequency}
\title{osm.getFrequency}
\usage{
osm.getFrequency(indf, inword = "shop", removeNA = T)
}
\arguments{
\item{indf}{The vec}

\item{inword}{Boole}

\item{removeNA}{Ad}
}
\value{
A nice histogram of a single variable with an additional Cumulative Density curve.
}
\description{
Get Frequency Table
}
\examples{
library(goal)

}
\seealso{
Other osm: 
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getRoads}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{POIs}
\keyword{frequency,}
\keyword{openstreetmap,}
