% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.r
\name{osm.getRoads}
\alias{osm.getRoads}
\title{osm.getRoads}
\usage{
osm.getRoads(
  incoordinates = c(26.545029, 39.088569, 26.570177, 39.11681),
  inPerioxi = "Mytilene Municipal Unit",
  withBB = FALSE,
  outcrs = 2100
)
}
\arguments{
\item{incoordinates}{Four coordinates of the bounding box.}

\item{inPerioxi}{Name of an area}

\item{withBB}{Boolean indicating if we shall use a bounding box (incoordinates) or the name of an area (inPerioxi).}

\item{outcrs}{CRS of the the output}
}
\value{
An sfnetwork
}
\description{
Get the roads of an area.
Either by name of area (inPerioxi) or by bounding box (incoordinates).
}
\examples{
library(goal)
library(sf)
library(sfnetworks)
library(osmdata)
mynetwork = osm.getRoads(incoordinates= c(26.545029,39.088569,26.570177,39.116810),
  withBB=TRUE, outcrs=2100
)
mynetwork

mynetwork2 = osm.getRoads(inPerioxi="Mytilene Municipal Unit",
  withBB=FALSE, outcrs=2100
)
mynetwork2

}
\seealso{
Other osm: 
\code{\link{osm.ClipSFnetwork_with_poly}()},
\code{\link{osm.CreateEmptyRaster}()},
\code{\link{osm.bb_2_pol}()},
\code{\link{osm.combineAmenities}()},
\code{\link{osm.combineShops}()},
\code{\link{osm.getFrequency}()},
\code{\link{osm.getLength_footway}()},
\code{\link{osm.getPOI}()},
\code{\link{osm.getPOI_usingbb}()},
\code{\link{osm.myBB_2_pol}()},
\code{\link{osm.osmdata_result_2_bbox_pol}()}
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\concept{osm}
\keyword{bbox,}
\keyword{network}
\keyword{openstreetmap,}
